Adding the Zoomable Circle Packing chart with iput file
